Move *_config to config packages
This way, we avoid having to pull every full library when linking with config.
