+++ src/libaktualizr/CMakeLists.txt
@@ -1,8 +1,10 @@
+# config has to go first, as other libraries append sources to it
+add_subdirectory("config")
+
 add_subdirectory("utilities")
 add_subdirectory("bootloader")
 add_subdirectory("bootstrap")
 add_subdirectory("campaign")
-add_subdirectory("config")
 add_subdirectory("crypto")
 add_subdirectory("http")
 add_subdirectory("primary")
@@ -31,8 +33,7 @@ add_library(aktualizr_static_lib STATIC
     $<TARGET_OBJECTS:logging>
     $<TARGET_OBJECTS:storage>
     $<TARGET_OBJECTS:uptane>
-    $<TARGET_OBJECTS:socket_activation>
-    $<TARGET_OBJECTS:telemetry>)
+    $<TARGET_OBJECTS:socket_activation>)
 
 if (BUILD_ISOTP)
     target_sources(aktualizr_static_lib PRIVATE $<TARGET_OBJECTS:isotp_conn>)
