+++ src/libaktualizr/package_manager/packagemanagerconfig.cc
@@ -2,24 +2,6 @@
 
 #include <boost/log/trivial.hpp>
 
-void PackageConfig::updateFromPropertyTree(const boost::property_tree::ptree& pt) {
-  CopyFromConfig(type, "type", pt);
-  CopyFromConfig(os, "os", pt);
-  CopyFromConfig(sysroot, "sysroot", pt);
-  CopyFromConfig(ostree_server, "ostree_server", pt);
-  CopyFromConfig(packages_file, "packages_file", pt);
-  CopyFromConfig(fake_need_reboot, "fake_need_reboot", pt);
-}
-
-void PackageConfig::writeToStream(std::ostream& out_stream) const {
-  writeOption(out_stream, type, "type");
-  writeOption(out_stream, os, "os");
-  writeOption(out_stream, sysroot, "sysroot");
-  writeOption(out_stream, ostree_server, "ostree_server");
-  writeOption(out_stream, packages_file, "packages_file");
-  writeOption(out_stream, fake_need_reboot, "fake_need_reboot");
-}
-
 std::ostream& operator<<(std::ostream& os, PackageManager pm) {
   std::string pm_str;
   switch (pm) {
@@ -36,4 +18,22 @@ std::ostream& operator<<(std::ostream& os, PackageManager pm) {
   }
   os << '"' << pm_str << '"';
   return os;
+}
+
+void PackageConfig::updateFromPropertyTree(const boost::property_tree::ptree& pt) {
+  CopyFromConfig(type, "type", pt);
+  CopyFromConfig(os, "os", pt);
+  CopyFromConfig(sysroot, "sysroot", pt);
+  CopyFromConfig(ostree_server, "ostree_server", pt);
+  CopyFromConfig(packages_file, "packages_file", pt);
+  CopyFromConfig(fake_need_reboot, "fake_need_reboot", pt);
+}
+
+void PackageConfig::writeToStream(std::ostream& out_stream) const {
+  writeOption(out_stream, type, "type");
+  writeOption(out_stream, os, "os");
+  writeOption(out_stream, sysroot, "sysroot");
+  writeOption(out_stream, ostree_server, "ostree_server");
+  writeOption(out_stream, packages_file, "packages_file");
+  writeOption(out_stream, fake_need_reboot, "fake_need_reboot");
 }
