+++ src/libaktualizr/storage/CMakeLists.txt
@@ -16,11 +16,9 @@ endif()
 set(HEADERS ${HEADERS} storage_config.h fsstorage_read.h invstorage.h)
 set(SOURCES ${SOURCES} fsstorage_read.cc invstorage.cc)
 
-aktualizr_source_file_checks(${SOURCES} ${HEADERS})
+target_sources(config PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/storage_config.cc)
 
 if(STORAGE_TYPE STREQUAL "sqlite")
-  list(APPEND SOURCES sql_schemas.cc)
-
   add_aktualizr_test(NAME storage_atomic SOURCES storage_atomic_test.cc PROJECT_WORKING_DIRECTORY)
   add_aktualizr_test(NAME sql_utils SOURCES sql_utils_test.cc PROJECT_WORKING_DIRECTORY)
   add_aktualizr_test(NAME sqlstorage SOURCES sqlstorage_test.cc sql_schemas.cc
@@ -33,6 +31,6 @@ if(STORAGE_TYPE STREQUAL "sqlite")
   set_tests_properties(test_schema_migration PROPERTIES LABELS "noptest")
 endif(STORAGE_TYPE STREQUAL "sqlite")
 
-add_library(storage OBJECT ${SOURCES})
+add_library(storage OBJECT ${SOURCES} sql_schemas.cc)
 
-aktualizr_source_file_checks(${TEST_SOURCES})
+aktualizr_source_file_checks(${SOURCES} ${HEADERS} storage_config.cc ${TEST_SOURCES})
