+++ src/libaktualizr/storage/storage_config.h
@@ -10,6 +10,7 @@
 #include "utilities/config_utils.h"
 
 enum class StorageType { kFileSystem = 0, kSqlite };
+std::ostream& operator<<(std::ostream& os, StorageType stype);
 
 struct StorageConfig {
   StorageType type{StorageType::kSqlite};
