+++ src/libaktualizr/telemetry/CMakeLists.txt
@@ -2,5 +2,5 @@ set(SOURCES telemetryconfig.cc)
 
 set(HEADERS telemetryconfig.h)
 
-add_library(telemetry OBJECT ${SOURCES})
+target_sources(config PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/telemetryconfig.cc)
 aktualizr_source_file_checks(${SOURCES} ${HEADERS})
